+++ spec/metadata/ios_parser.spec.js
@@ -22,28 +22,22 @@ var platforms = require('../../platforms'),
     shell = require('shelljs'),
     plist = require('plist-with-patches'),
     xcode = require('xcode'),
-    ET = require('elementtree'),
+    et = require('elementtree'),
     fs = require('fs'),
     Q = require('q'),
-    child_process = require('child_process'),
     config = require('../../src/config'),
-    config_parser = require('../../src/config_parser'),
+    ConfigParser = require('../../src/ConfigParser'),
     cordova = require('../../cordova');
 
 // Create a real config object before mocking out everything.
-var cfg = new config_parser(path.join(__dirname, '..', 'test-config.xml'));
+var cfg = new ConfigParser(path.join(__dirname, '..', 'test-config.xml'));
 
 describe('ios project parser', function () {
     var proj = path.join('some', 'path');
-    var exec, custom, readdir, cfg_parser;
+    var custom, readdir;
     beforeEach(function() {
-        exec = spyOn(child_process, 'exec').andCallFake(function(cmd, opts, cb) {
-            if (!cb) cb = opts;
-            cb(null, '', '');
-        });
         custom = spyOn(config, 'has_custom_path').andReturn(false);
         readdir = spyOn(fs, 'readdirSync').andReturn(['test.xcodeproj']);
-        cfg_parser = spyOn(util, 'config_parser');
     });
 
     function wrapper(p, done, post) {
@@ -89,29 +83,12 @@ describe('ios project parser', function () {
         });
 
         describe('update_from_config method', function() {
-            var et, xml, find, write_xml, root, mv;
-            var find_obj, root_obj, cfg_access_add, cfg_access_rm, cfg_pref_add, cfg_pref_rm, cfg_content;
+            var mv;
+            var cfg_access_add, cfg_access_rm, cfg_pref_add, cfg_pref_rm, cfg_content;
             var plist_parse, plist_build, xc;
             var update_name, xc_write;
             beforeEach(function() {
                 mv = spyOn(shell, 'mv');
-                find_obj = {
-                    text:'hi'
-                };
-                root_obj = {
-                    attrib:{
-                        package:'android_pkg'
-                    }
-                };
-                find = jasmine.createSpy('ElementTree find').andReturn(find_obj);
-                write_xml = jasmine.createSpy('ElementTree write');
-                root = jasmine.createSpy('ElementTree getroot').andReturn(root_obj);
-                et = spyOn(ET, 'ElementTree').andReturn({
-                    find:find,
-                    write:write_xml,
-                    getroot:root
-                });
-                xml = spyOn(ET, 'XML');
                 plist_parse = spyOn(plist, 'parseFileSync').andReturn({
                 });
                 plist_build = spyOn(plist, 'build').andReturn('');
@@ -125,27 +102,6 @@ describe('ios project parser', function () {
                 cfg.name = function() { return 'testname' };
                 cfg.packageName = function() { return 'testpkg' };
                 cfg.version = function() { return 'one point oh' };
-                cfg.access.get = function() { return [] };
-                cfg.preference.get = function() { return [] };
-                cfg.content = function() { return 'index.html'; };
-                cfg_access_add = jasmine.createSpy('config_parser access add');
-                cfg_access_rm = jasmine.createSpy('config_parser access rm');
-                cfg_pref_rm = jasmine.createSpy('config_parser pref rm');
-                cfg_pref_add = jasmine.createSpy('config_parser pref add');
-                cfg_content = jasmine.createSpy('config_parser content');
-                cfg_parser.andReturn({
-                    access:{
-                        remove:cfg_access_rm,
-                        get:function(){},
-                        add:cfg_access_add
-                    },
-                    preference:{
-                        remove:cfg_pref_rm,
-                        get:function(){},
-                        add:cfg_pref_add
-                    },
-                    content:cfg_content
-                });
                 p = new platforms.ios.parser(ios_proj);
             });
 
@@ -176,11 +132,6 @@ describe('ios project parser', function () {
                 expect(p.www_dir()).toEqual(path.join(ios_proj, 'www'));
             });
         });
-        describe('staging_dir method', function() {
-            it('should return .staging/www', function() {
-                expect(p.staging_dir()).toEqual(path.join(ios_proj, '.staging', 'www'));
-            });
-        });
         describe('config_xml method', function() {
             it('should return the location of the config.xml', function() {
                 expect(p.config_xml()).toEqual(path.join(ios_proj, 'test', 'config.xml'));
@@ -208,28 +159,12 @@ describe('ios project parser', function () {
                 expect(cp).toHaveBeenCalled();
             });
         });
-        describe('update_staging method', function() {
-            var exists;
-            beforeEach(function() {
-                exists = spyOn(fs, 'existsSync').andReturn(true);
-            });
-            it('should do nothing if staging dir does not exist', function() {
-                exists.andReturn(false);
-                p.update_staging();
-                expect(cp).not.toHaveBeenCalled();
-            });
-            it('should copy the staging dir into www if staging dir exists', function() {
-                p.update_staging();
-                expect(cp).toHaveBeenCalled();
-            });
-        });
         describe('update_project method', function() {
-            var config, www, overrides, staging, svn;
+            var config, www, overrides, svn;
             beforeEach(function() {
                 config = spyOn(p, 'update_from_config').andReturn(Q());
                 www = spyOn(p, 'update_www');
                 overrides = spyOn(p, 'update_overrides');
-                staging = spyOn(p, 'update_staging');
                 svn = spyOn(util, 'deleteSvnFolders');
             });
             it('should call update_from_config', function(done) {
@@ -252,11 +187,6 @@ describe('ios project parser', function () {
             it('should call update_overrides', function(done) {
                 wrapper(p.update_project(), done, function() {
                     expect(overrides).toHaveBeenCalled();
-                });
-            });
-            it('should call update_staging', function(done) {
-                wrapper(p.update_project(), done, function() {
-                    expect(staging).toHaveBeenCalled();
                 });
             });
             it('should call deleteSvnFolders', function(done) {