+++ RELEASENOTES.md
@@ -20,6 +20,15 @@
 -->
 # Cordova-cli Release Notes
 
+## 3.4.0-0.1.0 (Feb 14, 2014)
+* CB-5638 Clean-up: remove unreachable info case from function
+* CB-5937 Add "platform check" command: Shows platforms that are out of date
+* CB-5634 Minor refactoring + tests for Android's orientation preference.
+* CB-5634 Set Android orientation from config.xml
+* Upleveled amazon_fireos_parser. Making it at par with android_parser.js
+* CB-5947 Throw when trying to create project inside custom www.
+* CB-4153 Update help.txt about --source -> --copy-from
+
 ## 3.3.1-0.3.1 (Jan 31, 2014)
 * CB-4153 Rename --source and --link flags to --copy-from and --link-to
 
@@ -222,4 +231,3 @@ Important note: This version targets Cordova version 3.1.0-rc1.
 
 - Plugins are now installed serially across all installed platforms, rather than in parallel. This avoids race conditions in dependency installation. [CB-4184](https://issues.apache.org/jira/browse/CB-4184)
 - (WP8) All files from project www dir are now copied into the binary, not the top-level www. This means merges and plugin assets are correctly handled.
-